Corrections to support the new risc-v priviledge spec
Quite a few changes in this release, due to changes in QEMU that now
better fits to the official specs.
Mainly we enable all interrupts when creating a thread (kernel or user),
and we now acknowledge the timer interrupt in machine mode, as the
supervisor timer interrupt pending bit is read-only in supervisor mode.